Redesign TextProtocolBackendHandler  (#8493)
* Remove useless null judge

* Add SchemaBackendHandlerFactory

* Add SchemaBackendHandler

* Add environment package

* Move database related to data package

* Add data.impl package

* Add DatabaseBackendHandlerFactory

* Refactor TextProtocolBackendHandlerFactory

* Move RDLBackendHandler package

* Add default method for TextProtocolBackendHandler
