+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/java/org/apache/shardingsphere/proxy/backend/text/data/impl/UnicastDatabaseBackendHandler.java
@@ -15,26 +15,29 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.proxy.backend.text.metadata;
+package org.apache.shardingsphere.proxy.backend.text.data.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
+import org.apache.shardingsphere.proxy.backend.communication.DatabaseCommunicationEngine;
 import org.apache.shardingsphere.proxy.backend.communication.DatabaseCommunicationEngineFactory;
 import org.apache.shardingsphere.proxy.backend.communication.jdbc.connection.BackendConnection;
 import org.apache.shardingsphere.proxy.backend.context.ProxyContext;
+import org.apache.shardingsphere.proxy.backend.exception.NoDatabaseSelectedException;
 import org.apache.shardingsphere.proxy.backend.exception.RuleNotExistsException;
 import org.apache.shardingsphere.proxy.backend.response.header.ResponseHeader;
-import org.apache.shardingsphere.proxy.backend.response.header.update.UpdateResponseHeader;
-import org.apache.shardingsphere.proxy.backend.text.TextProtocolBackendHandler;
+import org.apache.shardingsphere.proxy.backend.text.data.DatabaseBackendHandler;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.SQLStatement;
 
 import java.sql.SQLException;
 import java.util.Collection;
+import java.util.Map;
 
 /**
- * Backend handler for broadcast.
+ * Database backend handler with unicast schema.
  */
 @RequiredArgsConstructor
-public final class BroadcastBackendHandler implements TextProtocolBackendHandler {
+public final class UnicastDatabaseBackendHandler implements DatabaseBackendHandler {
     
     private final DatabaseCommunicationEngineFactory databaseCommunicationEngineFactory = DatabaseCommunicationEngineFactory.getInstance();
     
@@ -44,27 +47,36 @@ public final class BroadcastBackendHandler implements TextProtocolBackendHandler
     
     private final BackendConnection backendConnection;
     
+    private DatabaseCommunicationEngine databaseCommunicationEngine;
+    
     @Override
     public ResponseHeader execute() throws SQLException {
-        String originalSchema = backendConnection.getSchemaName();
-        for (String each : ProxyContext.getInstance().getAllSchemaNames()) {
-            backendConnection.setCurrentSchema(each);
-            if (!ProxyContext.getInstance().getMetaData(each).isComplete()) {
-                throw new RuleNotExistsException();
-            }
-            databaseCommunicationEngineFactory.newTextProtocolInstance(sqlStatement, sql, backendConnection).execute();
+        if (null == backendConnection.getSchemaName()) {
+            // TODO should remove set default ShardingSphere schema after parser can recognize all DAL broadcast SQL.
+            backendConnection.setCurrentSchema(getFirstSchemaName());
         }
-        backendConnection.setCurrentSchema(originalSchema);
-        return new UpdateResponseHeader(sqlStatement);
+        if (!ProxyContext.getInstance().getMetaData(backendConnection.getSchemaName()).isComplete()) {
+            throw new RuleNotExistsException();
+        }
+        databaseCommunicationEngine = databaseCommunicationEngineFactory.newTextProtocolInstance(sqlStatement, sql, backendConnection);
+        return databaseCommunicationEngine.execute();
+    }
+    
+    private String getFirstSchemaName() {
+        Map<String, ShardingSphereMetaData> metaDataMap = ProxyContext.getInstance().getMetaDataContexts().getMetaDataMap();
+        if (metaDataMap.isEmpty()) {
+            throw new NoDatabaseSelectedException();
+        }
+        return metaDataMap.keySet().iterator().next();
     }
     
     @Override
-    public boolean next() {
-        return false;
+    public boolean next() throws SQLException {
+        return databaseCommunicationEngine.next();
     }
     
     @Override
-    public Collection<Object> getRowData() {
-        return null;
+    public Collection<Object> getRowData() throws SQLException {
+        return databaseCommunicationEngine.getQueryResponseRow().getData();
     }
 }
