+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/java/org/apache/shardingsphere/proxy/backend/text/skip/SkipBackendHandler.java
@@ -21,9 +21,6 @@ import org.apache.shardingsphere.proxy.backend.response.header.ResponseHeader;
 import org.apache.shardingsphere.proxy.backend.response.header.update.UpdateResponseHeader;
 import org.apache.shardingsphere.proxy.backend.text.TextProtocolBackendHandler;
 
-import java.util.Collection;
-import java.util.Collections;
-
 /**
  * Skip backend handler.
  */
@@ -32,15 +29,5 @@ public final class SkipBackendHandler implements TextProtocolBackendHandler {
     @Override
     public ResponseHeader execute() {
         return new UpdateResponseHeader(null);
-    }
-    
-    @Override
-    public boolean next() {
-        return false;
-    }
-    
-    @Override
-    public Collection<Object> getRowData() {
-        return Collections.emptyList();
     }
 }
